Refactor views for Create/Restore screens
