+++ app/src/main/res/color/stateful_button_yellow_color.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:color="?ColorJacobPressed" android:state_pressed="true" />
+    <item android:color="@color/yellow_50" android:state_enabled="false" />
     <item android:color="?ColorJacob" />
 </selector>
